% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_from_reference}
\alias{convert_from_reference}
\title{Convert one reference currency into all others}
\usage{
convert_from_reference(amount, ref_currency, currency_data)
}
\arguments{
\item{amount}{Numeric value. The amount to convert.}

\item{ref_currency}{Character. ISO code of the reference currency.}

\item{currency_data}{A tibble returned by load_currency_data().}
}
\value{
A tibble with columns: from, to, currency_name, converted_amount.
}
\description{
Convert one reference currency into all others
}
