% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_currency}
\alias{convert_currency}
\title{Convert an amount from one currency to another}
\usage{
convert_currency(amount, from, to, currency_data)
}
\arguments{
\item{amount}{Numeric value. The amount to convert.}

\item{from}{Character. ISO code of the source currency (e.g. "CHF").}

\item{to}{Character. ISO code of the target currency (e.g. "EUR").}

\item{currency_data}{A tibble returned by load_currency_data().
Contains at least iso_code and rate_to_usd.}
}
\value{
A numeric value corresponding to the converted amount.
}
\description{
Convert an amount from one currency to another
}
